refactor: move copy APIs out of util.go into parser_deep_copy.go
Addresses PR review feedback on file organization:
- Parser.DeepCopy/StructuralCopy (+WithTransform variants), the plan+fill
  copy machinery, heap-mode copy primitives, and the package-level copy
  API move verbatim from util.go to a new parser_deep_copy.go with an
  architecture overview comment.
- util.go keeps only generic helpers.
- transform_test.go is renamed to parser_transform_test.go since it
  exercises Parser methods.
